Refactor test item CLI to share pipeline helpers
